Adjust the > and < operators to act more like simple comparison
Instead of having these operators imply a !=.* this will instead
have them more-or-less simply check if the prospective version
is greater than or less than the specifier version. They will not
however match a post or pre-release of the version mentioned in
the specifier unless the specifier itself contains a post or
a pre-release.

In addition, we will ensure that >V does not match V+local.version
even though it is techincally greater than V.