Fix check for 64 bit OS
distutils.util.get_platform() returns "linux-x86_64" on 64 bit Linux and not
"linux_86_64" as assumed by this function. Instead we use the first element
returned by platform.architecture() and move the check into a separate fixture.

Furthermore, we have to check whether the current OS is x86-based, as the
results don't match otherwise.