Parse musl libc path from header
This is probably slightly faster and more robust than the ldd method
since it does not require ldd on PATH, and avoids some terminal encoding
issues.

The ldd approach is kept in edge cases where the executable is somehow
not readable. I suspect ldd would not be very useful in this scenario
either, but there's not harm being safe?