Fix failing parse action
pyparsing 3.0.5 renamed the two private attributes that are used
here.  This exact same transformation is performed by
`originalTextFor`'s default action, and assuming the intention
was simply to construct a `Marker` from the matched string, I've
restored the default action and added a secondary action for
constructing the `Marker` object.