Fix prefix version matching
This commit fixes prefix version matching when the prospective version contains a pre-release and/or post-release and/or dev segment.

This commit does not take into account pre-release and/or post-release segment in the specifier as PEP 440 is not precise enough for the behavior of such specifier and the existing code was not taking that into account either.