Reorder `Specifier` regex groups to check wildcard matches early
This ensures that an `re.match` or `re.search` does not exclude the
wildcard in cases where it would otherwise match. Without this
reordering, a regex match of a wildcard specifier would only match the
version sections before it since both parts of the prior part of the
unnamed group are optional, allowing the group to match an empty string
even though it is optional as a whole.

This does not affect `Specifier` since the regex used by it gets a `$`
added to it, forcing the regular expression engine to backtrack in such
cases. It does, however, affect `Requirement` since it is currently
allowing these values through since they match the regex for
`LegacySpecifier`.