Rework the `metadata` module annotations and documentation
The migration from modern annotations (thanks to a `__future__` import)
to the plain typing-import based annotations is to deal with limitations
of intersphinx with autodoc, which currently can't linkify annotations
with future style imports.

Similarly, to make the type annotations in the generated documentation
correctly link to the various "internal" types, they need to be imported
and referenced by their regular name instead of a changed-name
import. In the spirit of consistency, all imports were changed to match
this style.

This allows for eliminating duplication of content in the implementation
and documentation, reducing the general maintainance overhead of this
module.