Clarify version split/join usage (#725)
This adds a docstring on _version_join to clarify its expectation on the
input value. Since this is a strictly internal function, I also removed
the assert statement so it does not fire at runtime and break user code.
If it ever does fire, it would be due to a bug in packaging and we
should fix it instead.

Both _version_split and _version_join also received slight refactoring.