pylock: make from_dict a classmethod again
My idea was to have from_dict independent of the class
to cope for future evolution, but if a new version of the
standard can't be implemented as a subclass of Pylock,
return type of from_dict would change, so it would be
a breaking change.

If/when a new version of the spec arrive, it will still be
time to design the API evolution, depending of the actual changes.