Correct regex for metadata 'name' format (#925)
The current regex permits strings that end in the newline character, which is counter to what the description for the field states ("ASCII letters and numbers, period, underscore and hyphen").

This updates the rexex to use `\Z` instead of `$` to match at the end of the string and exclude newline characters.