chore: use typos instead of codespell
This is a Rust spell checker with the same basic idea as codespell, but it supports CamelCase / snake_case, and has a more flexible configuraiton (not perfect). I've found it catches some things codespell doesn't, and it's fast (though codespell is pretty fast too). It's got more stars on github than codespell, and has been around for a while, and seems well maintained. I haven't been actively trying it out that long, though (since the Python Bytes mention).

Signed-off-by: Henry Schreiner <henryschreineriii@gmail.com>